Correções no uso do hibernate

+++ src/main/java/ufg/pw/projeto_restaurante/model/cliente/dao/ClienteDao.java
@@ -3,7 +3,7 @@ package ufg.pw.projeto_restaurante.model.cliente.dao;
 import ufg.pw.projeto_restaurante.model.utils.persistencia.GenericDao;
 import ufg.pw.projeto_restaurante.model.cliente.Cliente;
 
-public class ClienteDao extends GenericDao<Cliente, Integer>{
+public class ClienteDao extends GenericDao<Cliente, Long>{
 	 public ClienteDao() {
 	       super(Cliente.class);
 	 }
